Upgrade to Node.js v14.17.6 when using Node Version Manager
We want to make sure our users are using a secure version of npm. This
commit bumps default Node.js version specified by `.nvmrc` to the latest
Node 14 version. This includes the latest version of npm, which is
patched to close an arbitrary code execution vulnerability. This follows
a recommendation by the GitHub security blog [[1]].

[1]: https://github.blog/2021-09-08-github-security-update-vulnerabilities-tar-npmcli-arborist/